remove the Assembly-CSharp stuff.
