+++ src/KKManager.Core/Data/Cards/AI/ChaFileParameter.cs
