+++ src/KKManager.Core/Data/Cards/KK/KoiCard.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.IO;
 using MessagePack;
-using KKManager.Data.Game;
 
 namespace KKManager.Data.Cards.KK
 {
