+++ src/KKManager.Core/packages.config
@@ -2,10 +2,6 @@
 <packages>
   <package id="Ben.Demystifier" version="0.4.1" targetFramework="net472" />
   <package id="DotNetZip" version="1.16.0" targetFramework="net472" />
-  <package id="Hjg.Pngcs" version="1.1.4" targetFramework="net472" />
-  <package id="Hst.Imaging" version="0.4.48" targetFramework="net472" />
-  <package id="Hst.Imaging.Pngcs" version="0.4.48" targetFramework="net472" />
-  <package id="libpng-sharp" version="0.1.18" targetFramework="net472" />
   <package id="MessagePack" version="2.5.108" targetFramework="net472" />
   <package id="MessagePack.Annotations" version="2.5.108" targetFramework="net472" />
   <package id="Microsoft.Bcl.AsyncInterfaces" version="7.0.0" targetFramework="net472" />
@@ -13,7 +9,6 @@
   <package id="Mono.Cecil" version="0.11.5" targetFramework="net472" />
   <package id="NetSettingBinder" version="1.1.6062.19744" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="13.0.3" targetFramework="net472" />
-  <package id="NUnit3TestAdapter" version="4.4.2" targetFramework="net472" />
   <package id="ObjectListView.Official" version="2.9.1" targetFramework="net472" />
   <package id="SharpCompress" version="0.33.0" targetFramework="net472" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
