+++ src/KKManager.Updater/packages.config
@@ -13,7 +13,6 @@
   <package id="Mono.Nat" version="3.0.4" targetFramework="net472" />
   <package id="MonoTorrent" version="3.0.0-beta.rev0159" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="13.0.3" targetFramework="net472" />
-  <package id="NUnit3TestAdapter" version="4.4.2" targetFramework="net472" />
   <package id="ObjectListView.Official" version="2.9.1" targetFramework="net472" />
   <package id="ReusableTasks" version="2.0.2" targetFramework="net472" />
   <package id="SharpCompress" version="0.33.0" targetFramework="net472" />
