+++ src/KKManager.sln
@@ -15,27 +15,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KKManager.Updater", "KKMana
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KKManager.SB3UGS", "KKManager.SB3UGS\KKManager.SB3UGS.csproj", "{B8B55E55-D95B-4E6A-9F33-C3737AD4487F}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{15B2FB54-E25D-4B98-B595-221AB9644273}"
-	ProjectSection(SolutionItems) = preProject
-		.editorconfig = .editorconfig
-	EndProjectSection
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CWrapperGen", "CWrapperGen\CWrapperGen.vcxproj", "{881B75AF-43B1-46BE-89ED-10A055E399C7}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "ProjectTests\Tests.csproj", "{82D7D7B6-C30C-4DF5-B99F-3955611AFDA9}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
 		Release|x64 = Release|x64
-		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{04E36698-5B58-4335-B4D6-218A0FA8B57E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{04E36698-5B58-4335-B4D6-218A0FA8B57E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{04E36698-5B58-4335-B4D6-218A0FA8B57E}.Debug|x64.ActiveCfg = Debug|x64
 		{04E36698-5B58-4335-B4D6-218A0FA8B57E}.Debug|x64.Build.0 = Debug|x64
 		{04E36698-5B58-4335-B4D6-218A0FA8B57E}.Debug|x86.ActiveCfg = Debug|x86
@@ -44,10 +29,6 @@ Global
 		{04E36698-5B58-4335-B4D6-218A0FA8B57E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{04E36698-5B58-4335-B4D6-218A0FA8B57E}.Release|x64.ActiveCfg = Release|x64
 		{04E36698-5B58-4335-B4D6-218A0FA8B57E}.Release|x64.Build.0 = Release|x64
-		{04E36698-5B58-4335-B4D6-218A0FA8B57E}.Release|x86.ActiveCfg = Release|Any CPU
-		{04E36698-5B58-4335-B4D6-218A0FA8B57E}.Release|x86.Build.0 = Release|Any CPU
-		{09FC4EF5-9694-4961-A136-DCFD8B5F1368}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{09FC4EF5-9694-4961-A136-DCFD8B5F1368}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{09FC4EF5-9694-4961-A136-DCFD8B5F1368}.Debug|x64.ActiveCfg = Debug|x64
 		{09FC4EF5-9694-4961-A136-DCFD8B5F1368}.Debug|x64.Build.0 = Debug|x64
 		{09FC4EF5-9694-4961-A136-DCFD8B5F1368}.Debug|x86.ActiveCfg = Debug|x86
@@ -56,10 +37,6 @@ Global
 		{09FC4EF5-9694-4961-A136-DCFD8B5F1368}.Release|Any CPU.Build.0 = Release|Any CPU
 		{09FC4EF5-9694-4961-A136-DCFD8B5F1368}.Release|x64.ActiveCfg = Release|x64
 		{09FC4EF5-9694-4961-A136-DCFD8B5F1368}.Release|x64.Build.0 = Release|x64
-		{09FC4EF5-9694-4961-A136-DCFD8B5F1368}.Release|x86.ActiveCfg = Release|Any CPU
-		{09FC4EF5-9694-4961-A136-DCFD8B5F1368}.Release|x86.Build.0 = Release|Any CPU
-		{AE2AF306-3A62-4EFE-982B-2E448B4C113A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AE2AF306-3A62-4EFE-982B-2E448B4C113A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AE2AF306-3A62-4EFE-982B-2E448B4C113A}.Debug|x64.ActiveCfg = Debug|x64
 		{AE2AF306-3A62-4EFE-982B-2E448B4C113A}.Debug|x64.Build.0 = Debug|x64
 		{AE2AF306-3A62-4EFE-982B-2E448B4C113A}.Debug|x86.ActiveCfg = Debug|x86
@@ -68,10 +45,6 @@ Global
 		{AE2AF306-3A62-4EFE-982B-2E448B4C113A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AE2AF306-3A62-4EFE-982B-2E448B4C113A}.Release|x64.ActiveCfg = Release|x64
 		{AE2AF306-3A62-4EFE-982B-2E448B4C113A}.Release|x64.Build.0 = Release|x64
-		{AE2AF306-3A62-4EFE-982B-2E448B4C113A}.Release|x86.ActiveCfg = Release|Any CPU
-		{AE2AF306-3A62-4EFE-982B-2E448B4C113A}.Release|x86.Build.0 = Release|Any CPU
-		{495A9EA6-B13C-48F9-A2E2-6667CECEBFDE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{495A9EA6-B13C-48F9-A2E2-6667CECEBFDE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{495A9EA6-B13C-48F9-A2E2-6667CECEBFDE}.Debug|x64.ActiveCfg = Debug|x64
 		{495A9EA6-B13C-48F9-A2E2-6667CECEBFDE}.Debug|x64.Build.0 = Debug|x64
 		{495A9EA6-B13C-48F9-A2E2-6667CECEBFDE}.Debug|x86.ActiveCfg = Debug|x86
@@ -80,10 +53,6 @@ Global
 		{495A9EA6-B13C-48F9-A2E2-6667CECEBFDE}.Release|Any CPU.Build.0 = Release|Any CPU
 		{495A9EA6-B13C-48F9-A2E2-6667CECEBFDE}.Release|x64.ActiveCfg = Release|x64
 		{495A9EA6-B13C-48F9-A2E2-6667CECEBFDE}.Release|x64.Build.0 = Release|x64
-		{495A9EA6-B13C-48F9-A2E2-6667CECEBFDE}.Release|x86.ActiveCfg = Release|Any CPU
-		{495A9EA6-B13C-48F9-A2E2-6667CECEBFDE}.Release|x86.Build.0 = Release|Any CPU
-		{B64C974D-9687-4E69-AEEA-3F04F769E556}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B64C974D-9687-4E69-AEEA-3F04F769E556}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B64C974D-9687-4E69-AEEA-3F04F769E556}.Debug|x64.ActiveCfg = Debug|x64
 		{B64C974D-9687-4E69-AEEA-3F04F769E556}.Debug|x64.Build.0 = Debug|x64
 		{B64C974D-9687-4E69-AEEA-3F04F769E556}.Debug|x86.ActiveCfg = Debug|x86
@@ -92,10 +61,6 @@ Global
 		{B64C974D-9687-4E69-AEEA-3F04F769E556}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B64C974D-9687-4E69-AEEA-3F04F769E556}.Release|x64.ActiveCfg = Release|x64
 		{B64C974D-9687-4E69-AEEA-3F04F769E556}.Release|x64.Build.0 = Release|x64
-		{B64C974D-9687-4E69-AEEA-3F04F769E556}.Release|x86.ActiveCfg = Release|Any CPU
-		{B64C974D-9687-4E69-AEEA-3F04F769E556}.Release|x86.Build.0 = Release|Any CPU
-		{B8B55E55-D95B-4E6A-9F33-C3737AD4487F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B8B55E55-D95B-4E6A-9F33-C3737AD4487F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B8B55E55-D95B-4E6A-9F33-C3737AD4487F}.Debug|x64.ActiveCfg = Debug|x64
 		{B8B55E55-D95B-4E6A-9F33-C3737AD4487F}.Debug|x64.Build.0 = Debug|x64
 		{B8B55E55-D95B-4E6A-9F33-C3737AD4487F}.Debug|x86.ActiveCfg = Debug|x86
