+++ src/KKManager/packages.config
@@ -12,7 +12,6 @@
   <package id="NBug" version="1.2.2" targetFramework="net472" />
   <package id="NetSettingBinder" version="1.1.6062.19744" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="13.0.3" targetFramework="net472" />
-  <package id="NUnit3TestAdapter" version="4.4.2" targetFramework="net472" />
   <package id="ObjectListView.Official" version="2.9.1" targetFramework="net472" />
   <package id="SharpCompress" version="0.33.0" targetFramework="net472" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
