--- src/PortableSettingsProvider/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NUnit3TestAdapter" version="4.4.2" targetFramework="net472" />
-</packages>
