+++ src/StandaloneUpdater/packages.config
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="NBug" version="1.2.2" targetFramework="net472" />
-  <package id="NUnit3TestAdapter" version="4.4.2" targetFramework="net472" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
   <package id="System.Collections.Immutable" version="7.0.0" targetFramework="net472" />
   <package id="System.Memory" version="4.5.5" targetFramework="net472" />
